debug and make postman see result

+++ src/utils/cloudinary.js
@@ -1,4 +1,4 @@
-import { v2 } from "cloudinary";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs"
 
 
@@ -9,23 +9,30 @@ cloudinary.config({
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if (!localFilePath) return null
-        //upload the file on cloudinary
-        cloudinary.uploader.upload(localFilePath,{
-            resource_type: "auto"
-        })
-        // file has been uploaded successfull
-        console.log("file is uploaded on cloudinary",
-            response.url);
-            return response;
-    } catch (error) {
-        fs.unlinkSync(localFilePath)
-         // remove the locally saved temporary file as the upload operation got failed
-        return null;
-    }
-}
+  try {
+    if (!localFilePath) return null;
 
+    // upload the file on cloudinary
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
+
+    console.log("file is uploaded on cloudinary:", response.url);
+
+    // delete local file after upload
+    fs.unlinkSync(localFilePath);
+
+    return response;
+  } catch (error) {
+    console.error("Cloudinary error:", error);
+
+    if (localFilePath) {
+      fs.unlinkSync(localFilePath);
+    }
+
+    return null;
+  }
+};
 
 export {uploadOnCloudinary}
